% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqi_package.R
\docType{package}
\name{cqi}
\alias{cqi}
\alias{cqi-package}
\title{cqi}
\description{
cqi
}
\examples{
\dontrun{
cqpserver -I /Library/Frameworks/R.framework/Versions/3.2/Resources/library/cqi/init/cqpserver.init -r /Users/blaette/Lab/cwb/registry
}
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\references{
http://cwb.sourceforge.net/cqi.php
}
\keyword{package}

